requirements.txt
mysql-connector-python==8.0.32
python-dotenv==1.0.0
bcrypt==4.0.1
colorama==0.4.6
tabulate==0.9.0 